% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_child_nebula.R
\name{annotate_child_nebulae}
\alias{annotate_child_nebulae}
\title{annotate_child_nebulae}
\usage{
annotate_child_nebulae(
  nebula_name,
  compound_class_list = .MCn.nebula_class,
  write_output = T,
  output = paste0(.MCn.output, "/", .MCn.results),
  layout = "fr",
  height = "auto",
  width = "auto",
  plot_nodes_id = T,
  plot_structure = T,
  plot_ppcp = T,
  ratio_df = NA,
  merge_image = T,
  return_plot = F,
  nodes_mark = NA,
  global.node.size = 0.6,
  theme_args = NA,
  ...
)
}
\arguments{
\item{nebula_name}{Character. The name of child-nebula.}

\item{compound_class_list}{A list, generaged by \code{collate_ppcp}, Default: .MCn.nebula_class}

\item{write_output}{Logic. Write output to a directory, Default: T}

\item{output}{Character, Default: paste0(.MCn.output, "/", .MCn.results)}

\item{layout}{Character, 'igraph' layout, Default: 'fr'}

\item{height}{A number, Default: 'auto'}

\item{width}{A number, Default: 'auto'}

\item{plot_nodes_id}{Logic, Default: T}

\item{plot_structure}{Logic, Default: T}

\item{plot_ppcp}{Logic, Default: T}

\item{ratio_df}{A data.frame, Default: NA}

\item{merge_image}{Logic, Default: T}

\item{return_plot}{Logic, Default: F}

\item{nodes_mark}{A is.data.frame, Default: NA}

\item{global.node.size}{A number, Default: 0.6}

\item{theme_args}{A list, Default: NA}

\item{...}{...}
}
\description{
Visualization of structure, PPCP and statistic data in child-nebula.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate}}
\code{\link[data.table]{rbindlist}}
}
