% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_mode_structure.R
\name{batch_mode_structure}
\alias{batch_mode_structure}
\title{FUNCTION_TITLE}
\usage{
batch_mode_structure(metadata, tmp_stru)
}
\arguments{
\item{metadata}{PARAM_DESCRIPTION}

\item{tmp_stru}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}
\code{\link[pbapply]{pbapply}}
}
