% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_classes_tree_list.R
\name{build_classes_tree_list}
\alias{build_classes_tree_list}
\title{build_classes_tree_list}
\usage{
build_classes_tree_list(class_index = "canopus.tsv", path = .MCn.sirius)
}
\arguments{
\item{class_index}{Character, source data name, Default: 'canopus.tsv'}

\item{path}{SIRIUS project path, Default: .MCn.sirius}
}
\description{
According to source data (table) of classification in SIRIUS project to build list in hierarchy.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{reexports}}
}
