% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_ppcp.R
\name{collate_ppcp}
\alias{collate_ppcp}
\title{FUNCTION_TITLE}
\usage{
collate_ppcp(
  dirs = "all",
  write_output = T,
  nebula_class = T,
  nebula_index = T,
  ...
)
}
\arguments{
\item{dirs}{PARAM_DESCRIPTION, Default: 'all'}

\item{write_output}{PARAM_DESCRIPTION, Default: T}

\item{nebula_class}{PARAM_DESCRIPTION, Default: T}

\item{nebula_index}{PARAM_DESCRIPTION, Default: T}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pbapply]{pbapply}}
\code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{reexports}}
\code{\link[data.table]{rbindlist}}
}
