% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_structure.R
\name{collate_structure}
\alias{collate_structure}
\title{FUNCTION_TITLE}
\usage{
collate_structure(dirs = "all", write_output = T, ...)
}
\arguments{
\item{dirs}{PARAM_DESCRIPTION, Default: 'all'}

\item{write_output}{PARAM_DESCRIPTION, Default: T}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pbapply]{pbapply}}
\code{\link[data.table]{rbindlist}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{filter}}
}
