% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_child_nebulae.R
\name{generate_child_nebulae}
\alias{generate_child_nebulae}
\title{generate_child_nebulae}
\usage{
generate_child_nebulae(
  nodes = .MCn.parent_nodes,
  edges = .MCn.parent_edges,
  max_edges = 5,
  nebula_index = .MCn.nebula_index,
  output = paste0(.MCn.output, "/", .MCn.results),
  output_format = "graphml",
  ...
)
}
\arguments{
\item{nodes}{A data.frame, Default: .MCn.parent_nodes}

\item{edges}{A data.frame, Default: .MCn.parent_edges}

\item{max_edges}{A number, Default: 5}

\item{nebula_index}{A data.frame, Default: .MCn.nebula_index}

\item{output}{Character, Default: paste0(.MCn.output, "/", .MCn.results)}

\item{output_format}{Character, 'igraph' supported format, Default: 'graphml'}

\item{...}{...}
}
\description{
According to nodes and edges data of parent-nebula and nebula-index to generate child-nebula.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pbapply]{pbapply}}
}
