% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parent_nebula.R
\name{generate_parent_nebula}
\alias{generate_parent_nebula}
\title{generate_parent_nebula}
\usage{
generate_parent_nebula(
  write_output = T,
  output_format = "graphml",
  output = paste0(.MCn.output, "/", .MCn.results),
  edges_file = paste0(output, "/parent_nebula/parent_nebula_edges.tsv"),
  edges_method = "method_formula_based_spec_compare",
  nodes_attributes = .MCn.formula_set,
  nodes_other_attributes = .MCn.structure_set,
  rm_parent_isolate_nodes = F,
  edge_filter = 0.5,
  cpu_cores = 8,
  ...
)
}
\arguments{
\item{write_output}{Logic, Default: T}

\item{output_format}{Character, 'igraph' supported format, Default: 'graphml'}

\item{output}{Character, Default: paste0(.MCn.output, "/", .MCn.results)}

\item{edges_file}{Character, path to edges file, Default: paste0(output, "/parent_nebula/parent_nebula_edges.tsv")}

\item{edges_method}{Character, Default: 'method_formula_based_spec_compare'}

\item{nodes_attributes}{A data.frame, formula dataset, Default: .MCn.formula_set}

\item{nodes_other_attributes}{A data.frame, structure dataset, Default: .MCn.structure_set}

\item{rm_parent_isolate_nodes}{Logic, Default: F}

\item{edge_filter}{A number, Default: 0.5}

\item{cpu_cores}{A number, thread count, Default: 8}

\item{...}{...}
}
\description{
According to formula dataset and structure dataset to generate parent-nebula graph
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{reexports}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{filter}}
\code{\link[igraph]{as_data_frame}}, \code{\link[igraph]{write_graph}}
}
