% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formula.R
\name{get_formula}
\alias{get_formula}
\title{get_formula}
\usage{
get_formula(
  key_id,
  exclude_element = NULL,
  formula_method = "top_zodiac",
  rank = 1:5,
  ppm_error = 20,
  return_col = c("rank", "precursorFormula", "molecularFormula", "adduct",
    "ZodiacScore", "massErrorPrecursor(ppm)"),
  ...
)
}
\arguments{
\item{key_id}{Character.}

\item{exclude_element}{Vector, Default: NULL}

\item{formula_method}{Character, Default: 'top_zodiac'}

\item{rank}{Vector of number, Default: 1:5}

\item{ppm_error}{A bumber, Default: 20}

\item{return_col}{Vector of character, Default: c("rank", "precursorFormula",
"massErrorPrecursor(ppm)")}

\item{...}{...}
}
\value{
A data.frame
}
\description{
A function to read table of formula data of feature in SIRIUS project.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
