% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppcp.R
\name{get_ppcp}
\alias{get_ppcp}
\title{FUNCTION_TITLE}
\usage{
get_ppcp(
  key_id = NULL,
  dir = NULL,
  precursor_formula = "method_pick_formula_excellent",
  adduct = NULL,
  reformat = T,
  filter = T,
  filter_threshold = 0.1,
  class_index = "canopus.tsv",
  ...
)
}
\arguments{
\item{key_id}{PARAM_DESCRIPTION, Default: NULL}

\item{dir}{PARAM_DESCRIPTION, Default: NULL}

\item{precursor_formula}{PARAM_DESCRIPTION, Default: 'method_pick_formula_excellent'}

\item{adduct}{PARAM_DESCRIPTION, Default: NULL}

\item{reformat}{PARAM_DESCRIPTION, Default: T}

\item{filter}{PARAM_DESCRIPTION, Default: T}

\item{filter_threshold}{PARAM_DESCRIPTION, Default: 0.1}

\item{class_index}{PARAM_DESCRIPTION, Default: 'canopus.tsv'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
