% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure.R
\name{get_structure}
\alias{get_structure}
\title{FUNCTION_TITLE}
\usage{
get_structure(
  key_id,
  precursor_formula = NULL,
  adduct = NULL,
  structure_method = "top_score",
  order = T,
  return_row = 1:10,
  as_tibble = F,
  ...
)
}
\arguments{
\item{key_id}{PARAM_DESCRIPTION}

\item{precursor_formula}{PARAM_DESCRIPTION, Default: NULL}

\item{adduct}{PARAM_DESCRIPTION, Default: NULL}

\item{structure_method}{PARAM_DESCRIPTION, Default: 'top_score'}

\item{order}{PARAM_DESCRIPTION, Default: T}

\item{return_row}{PARAM_DESCRIPTION, Default: 1:10}

\item{as_tibble}{PARAM_DESCRIPTION, Default: F}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{rbindlist}}
\code{\link[dplyr]{reexports}}
}
