% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_mcnebula.R
\name{initialize_mcnebula}
\alias{initialize_mcnebula}
\title{FUNCTION_TITLE}
\usage{
initialize_mcnebula(
  sirius_path,
  output_path = sirius_path,
  output_file = "mcnebula_results",
  palette = unique(c((ggsci::pal_simpsons())(16), (ggsci::pal_igv("default"))(51),
    (ggsci::pal_ucscgb())(26), (ggsci::pal_d3("category20"))(20))),
  palette_stat = palette,
  palette_ppcp = palette,
  palette_label = colorRampPalette(c("#C6DBEFFF", "#3182BDFF", "red"))(10),
  rm_var = F
)
}
\arguments{
\item{sirius_path}{PARAM_DESCRIPTION}

\item{output_path}{PARAM_DESCRIPTION, Default: sirius_path}

\item{output_file}{PARAM_DESCRIPTION, Default: 'mcnebula_results'}

\item{palette}{PARAM_DESCRIPTION, Default: unique(c((ggsci::pal_simpsons())(16), (ggsci::pal_igv("default"))(51)))}

\item{palette_stat}{PARAM_DESCRIPTION, Default: palette}

\item{palette_label}{PARAM_DESCRIPTION, Default: colorRampPalette(c("#C6DBEFFF", "#3182BDFF", "red"))(10)}

\item{rm_var}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggsci]{pal_simpsons}}, \code{\link[ggsci]{pal_igv}}
}
