% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_formula_based_spec_compare.R
\name{method_formula_based_spec_compare}
\alias{method_formula_based_spec_compare}
\title{FUNCTION_TITLE}
\usage{
method_formula_based_spec_compare(
  path = .MCn.sirius,
  dirs = "all",
  cpu_cores = 8,
  compare_fun = "dotproduct",
  precursor_mass_diff = T,
  edge_filter = 0.3,
  only_identical_class = T,
  min_hierarchy = 5,
  filter_only_max = 2000,
  min_zodiac = 0.9,
  min_tanimoto = 0.4,
  ...
)
}
\arguments{
\item{path}{PARAM_DESCRIPTION, Default: .MCn.sirius}

\item{dirs}{PARAM_DESCRIPTION, Default: 'all'}

\item{cpu_cores}{PARAM_DESCRIPTION, Default: 8}

\item{compare_fun}{PARAM_DESCRIPTION, Default: 'dotproduct'}

\item{precursor_mass_diff}{PARAM_DESCRIPTION, Default: T}

\item{edge_filter}{PARAM_DESCRIPTION, Default: 0.3}

\item{only_identical_class}{PARAM_DESCRIPTION, Default: T}

\item{min_hierarchy}{PARAM_DESCRIPTION, Default: 5}

\item{filter_only_max}{PARAM_DESCRIPTION, Default: 2000}

\item{min_zodiac}{PARAM_DESCRIPTION, Default: 0.9}

\item{min_tanimoto}{PARAM_DESCRIPTION, Default: 0.4}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pbapply]{pbapply}}
\code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise_all}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{reexports}}
}
