% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_pick_formula_excellent.R
\name{method_pick_formula_excellent}
\alias{method_pick_formula_excellent}
\title{FUNCTION_TITLE}
\usage{
method_pick_formula_excellent(
  key_id = NULL,
  dir = NULL,
  exclude_element = NULL,
  ppm_error = 20,
  fc = NA
)
}
\arguments{
\item{key_id}{PARAM_DESCRIPTION, Default: NULL}

\item{dir}{PARAM_DESCRIPTION, Default: NULL}

\item{exclude_element}{PARAM_DESCRIPTION, Default: NULL}

\item{ppm_error}{PARAM_DESCRIPTION, Default: 20}

\item{fc}{PARAM_DESCRIPTION, Default: 1.5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
