% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_summarize_nebula_class.R
\name{method_summarize_nebula_class}
\alias{method_summarize_nebula_class}
\title{FUNCTION_TITLE}
\usage{
method_summarize_nebula_class(
  data,
  ppcp_threshold = 0.5,
  max_classes = NA,
  hierarchy_priority = c(6, 5, 4, 3),
  class_data_type = "classes_tree_list",
  ...
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{ppcp_threshold}{PARAM_DESCRIPTION, Default: 0.5}

\item{max_classes}{PARAM_DESCRIPTION, Default: 5}

\item{hierarchy_priority}{PARAM_DESCRIPTION, Default: c(6, 5, 4, 3)}

\item{class_data_type}{PARAM_DESCRIPTION, Default: 'classes_tree_list'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{rbindlist}}
\code{\link[dplyr]{rename}}, \code{\link[dplyr]{filter}}
}
