% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_summarize_nebula_index.R
\name{method_summarize_nebula_index}
\alias{method_summarize_nebula_index}
\title{FUNCTION_TITLE}
\usage{
method_summarize_nebula_index(
  ppcp_dataset,
  nebula_class = .MCn.nebula_class,
  ppcp_threshold = 0.5,
  min_possess = 10,
  max_possess_pct = 0.2,
  filter_identical = c(top_hierarchy = 4),
  identical_factor = 0.7,
  rm_position_describe_class = T,
  filter_via_struc_score = "tanimotoSimilarity",
  struc_score_cutoff = 0.3,
  min_reached_pct = 0.6,
  ...
)
}
\arguments{
\item{ppcp_dataset}{PARAM_DESCRIPTION}

\item{nebula_class}{PARAM_DESCRIPTION, Default: .MCn.nebula_class}

\item{ppcp_threshold}{PARAM_DESCRIPTION, Default: 0.5}

\item{min_possess}{PARAM_DESCRIPTION, Default: 10}

\item{max_possess_pct}{PARAM_DESCRIPTION, Default: 0.2}

\item{filter_identical}{PARAM_DESCRIPTION, Default: c(top_hierarchy = 4)}

\item{identical_factor}{PARAM_DESCRIPTION, Default: 0.8}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{rbindlist}}
\code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{group_by}}
\code{\link[pbapply]{pbapply}}
}
