% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_child_nebulae.R
\name{visualize_child_nebulae}
\alias{visualize_child_nebulae}
\title{FUNCTION_TITLE}
\usage{
visualize_child_nebulae(
  graph_list = .MCn.child_graph_list,
  compound_class_list = .MCn.nebula_class,
  output = paste0(.MCn.output, "/", .MCn.results),
  layout = "fr",
  width = 23,
  height = 30,
  nodes_mark = NA,
  ...
)
}
\arguments{
\item{graph_list}{PARAM_DESCRIPTION, Default: .MCn.child_graph_list}

\item{compound_class_list}{PARAM_DESCRIPTION, Default: .MCn.nebula_class}

\item{output}{PARAM_DESCRIPTION, Default: paste0(.MCn.output, "/", .MCn.results)}

\item{layout}{PARAM_DESCRIPTION, Default: 'fr'}

\item{width}{PARAM_DESCRIPTION, Default: 23}

\item{height}{PARAM_DESCRIPTION, Default: 30}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{rbindlist}}
\code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{mutate}}
\code{\link[svglite]{svglite}}
\code{\link[grid]{grid.newpage}}, \code{\link[grid]{Working with Viewports}}
\code{\link[pbapply]{pbapply}}
}
