% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_nebula.R
\name{visualize_nebula}
\alias{visualize_nebula}
\title{FUNCTION_TITLE}
\usage{
visualize_nebula(nebula_name, ...)
}
\arguments{
\item{nebula_name}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
