% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_parent_nebula.R
\name{visualize_parent_nebula}
\alias{visualize_parent_nebula}
\title{FUNCTION_TITLE}
\usage{
visualize_parent_nebula(
  graph = .MCn.parent_graph,
  write_output = T,
  output = paste0(.MCn.output, "/", .MCn.results),
  layout = "mds",
  nodes_color = c(hierarchy = 3),
  width = 15,
  height = 12,
  return_plot = F,
  ...
)
}
\arguments{
\item{graph}{PARAM_DESCRIPTION, Default: .MCn.parent_graph}

\item{write_output}{PARAM_DESCRIPTION, Default: T}

\item{output}{PARAM_DESCRIPTION, Default: paste0(.MCn.output, "/", .MCn.results)}

\item{layout}{PARAM_DESCRIPTION, Default: 'mds'}

\item{nodes_color}{PARAM_DESCRIPTION, Default: c(hierarchy = 4)}

\item{width}{PARAM_DESCRIPTION, Default: 15}

\item{height}{PARAM_DESCRIPTION, Default: 12}

\item{return_plot}{PARAM_DESCRIPTION, Default: F}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pbapply]{pbapply}}
\code{\link[data.table]{rbindlist}}
\code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{reexports}}
\code{\link[tidygraph]{as_tbl_graph.data.frame}}, \code{\link[tidygraph]{activate}}
\code{\link[ggraph]{ggraph}}
\code{\link[ggplot2]{ggsave}}
}
