% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_child_nebula.R
\name{annotate_child_nebulae}
\alias{annotate_child_nebulae}
\title{FUNCTION_TITLE}
\usage{
annotate_child_nebulae(
  nebula_name,
  compound_class_list = .MCn.nebula_class,
  write_output = T,
  output = paste0(.MCn.output, "/", .MCn.results),
  layout = "fr",
  height = "auto",
  width = "auto",
  plot_nodes_id = T,
  plot_structure = T,
  plot_ppcp = T,
  ratio_df = NA,
  merge_image = T,
  return_plot = F,
  nodes_mark = NA,
  global.node.size = 0.6,
  theme_args = NA,
  ...
)
}
\arguments{
\item{nebula_name}{PARAM_DESCRIPTION}

\item{compound_class_list}{PARAM_DESCRIPTION, Default: .MCn.nebula_class}

\item{write_output}{PARAM_DESCRIPTION, Default: T}

\item{output}{PARAM_DESCRIPTION, Default: paste0(.MCn.output, "/", .MCn.results)}

\item{layout}{PARAM_DESCRIPTION, Default: 'fr'}

\item{height}{PARAM_DESCRIPTION, Default: 'auto'}

\item{width}{PARAM_DESCRIPTION, Default: 'auto'}

\item{plot_nodes_id}{PARAM_DESCRIPTION, Default: T}

\item{plot_structure}{PARAM_DESCRIPTION, Default: T}

\item{plot_ppcp}{PARAM_DESCRIPTION, Default: T}

\item{ratio_df}{PARAM_DESCRIPTION, Default: NULL}

\item{merge_image}{PARAM_DESCRIPTION, Default: T}

\item{return_plot}{PARAM_DESCRIPTION, Default: F}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}
\code{\link[data.table]{rbindlist}}
\code{\link[ggraph]{geom_node_text}}
\code{\link[ggplot2]{c("guide_bins", "guide_colourbar", "guide_coloursteps", "guide_legend", "guides", "guides")}}, \code{\link[ggplot2]{ggsave}}
}
