% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_classes_tree_list.R
\name{build_classes_tree_list}
\alias{build_classes_tree_list}
\title{FUNCTION_TITLE}
\usage{
build_classes_tree_list(class_index = "canopus.tsv", path = .MCn.sirius)
}
\arguments{
\item{class_index}{PARAM_DESCRIPTION, Default: 'canopus.tsv'}

\item{path}{PARAM_DESCRIPTION, Default: .MCn.sirius}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{reexports}}
}
