% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_ppcp.R
\name{collate_ppcp}
\alias{collate_ppcp}
\title{collate_ppcp}
\usage{
collate_ppcp(
  dirs = "all",
  write_output = T,
  nebula_class = T,
  nebula_index = T,
  ...
)
}
\arguments{
\item{dirs}{Vector, Default: 'all'}

\item{write_output}{Logic, Default: T}

\item{nebula_class}{Logic, Default: T}

\item{nebula_index}{Logic, Default: T}

\item{...}{...}
}
\description{
Collate posterior probability of classification prediction (PPCP) from SIRIUS project
and conduct integration to get nebula_class and nebula-index.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pbapply]{pbapply}}
\code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{reexports}}
\code{\link[data.table]{rbindlist}}
}
