% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_structure.R
\name{collate_structure}
\alias{collate_structure}
\title{collate_structure}
\usage{
collate_structure(dirs = "all", write_output = T, ...)
}
\arguments{
\item{dirs}{Vector, Default: 'all'}

\item{write_output}{Logic, Default: T}

\item{...}{...}
}
\description{
Collate chemical structure data from SIRIUS project
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pbapply]{pbapply}}
\code{\link[data.table]{rbindlist}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{reexports}}
}
