% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_child_nebulae.R
\name{generate_child_nebulae}
\alias{generate_child_nebulae}
\title{FUNCTION_TITLE}
\usage{
generate_child_nebulae(
  nodes = .MCn.parent_nodes,
  edges = .MCn.parent_edges,
  max_edges = 5,
  nebula_index = .MCn.nebula_index,
  output = paste0(.MCn.output, "/", .MCn.results),
  output_format = "graphml",
  ...
)
}
\arguments{
\item{nodes}{PARAM_DESCRIPTION, Default: .MCn.parent_nodes}

\item{edges}{PARAM_DESCRIPTION, Default: .MCn.parent_edges}

\item{max_edges}{PARAM_DESCRIPTION, Default: 5}

\item{nebula_index}{PARAM_DESCRIPTION, Default: .MCn.nebula_index}

\item{output}{PARAM_DESCRIPTION, Default: paste0(.MCn.output, "/", .MCn.results)}

\item{output_format}{PARAM_DESCRIPTION, Default: 'graphml'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pbapply]{pbapply}}
}
