% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parent_nebula.R
\name{generate_parent_nebula}
\alias{generate_parent_nebula}
\title{FUNCTION_TITLE}
\usage{
generate_parent_nebula(
  write_output = T,
  output_format = "graphml",
  output = paste0(.MCn.output, "/", .MCn.results),
  edges_file = paste0(output, "/parent_nebula/parent_nebula_edges.tsv"),
  edges_method = "method_formula_based_spec_compare",
  nodes_attributes = .MCn.formula_set,
  nodes_other_attributes = .MCn.structure_set,
  rm_parent_isolate_nodes = F,
  edge_filter = 0.5,
  cpu_cores = 8,
  ...
)
}
\arguments{
\item{write_output}{PARAM_DESCRIPTION, Default: T}

\item{output_format}{PARAM_DESCRIPTION, Default: 'graphml'}

\item{output}{PARAM_DESCRIPTION, Default: paste0(.MCn.output, "/", .MCn.results)}

\item{edges_file}{PARAM_DESCRIPTION, Default: paste0(output, "/parent_nebula/parent_nebula_edges.tsv")}

\item{edges_method}{PARAM_DESCRIPTION, Default: 'method_formula_based_spec_compare'}

\item{nodes_attributes}{PARAM_DESCRIPTION, Default: .MCn.formula_set}

\item{nodes_other_attributes}{PARAM_DESCRIPTION, Default: .MCn.structure_set}

\item{edge_filter}{PARAM_DESCRIPTION, Default: 0.5}

\item{cpu_cores}{PARAM_DESCRIPTION, Default: 8}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{reexports}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr]{rename}}
\code{\link[igraph]{as_data_frame}}
}
