% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formula.R
\name{get_formula}
\alias{get_formula}
\title{FUNCTION_TITLE}
\usage{
get_formula(
  key_id,
  exclude_element = NULL,
  formula_method = "top_zodiac",
  rank = 1:5,
  ppm_error = 20,
  return_col = c("rank", "precursorFormula", "molecularFormula", "adduct",
    "ZodiacScore", "massErrorPrecursor(ppm)"),
  ...
)
}
\arguments{
\item{key_id}{PARAM_DESCRIPTION}

\item{exclude_element}{PARAM_DESCRIPTION, Default: NULL}

\item{formula_method}{PARAM_DESCRIPTION, Default: 'top_zodiac'}

\item{rank}{PARAM_DESCRIPTION, Default: 1:5}

\item{ppm_error}{PARAM_DESCRIPTION, Default: 20}

\item{return_col}{PARAM_DESCRIPTION, Default: c("rank", "precursorFormula", "molecularFormula", "adduct", "ZodiacScore",
"massErrorPrecursor(ppm)")}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
